feat: add useGetMyUser.ts
