+++ 2024-summer-FE-seminar/packages/api/src/drizzle/schema/promotional-printing.schema.ts
@@ -12,24 +12,25 @@ import {
 import { Club } from "./club.schema";
 import { Student } from "./user.schema";
 
-/** 
- * # HW1
- * DB 설계에 맞추어 drizzle schema를 구현해 봅시다.
- * 진행 중 아무리 검색하고 고민해 보아도 해결되지 않는 점이 있다면,
- * `promotional-printing.schema copy.ts` 파일을 참고하세요.
- */
-
 export const PromotionalPrintingOrderStatusEnum = mysqlTable(
   "promotional_printing_order_status_enum",
   {
-
+    id: int("id").autoincrement().primaryKey(),
+    statusName: varchar("status_name", { length: 30 }),
+    createdAt: timestamp("created_at").defaultNow().notNull(),
+    updatedAt: timestamp("updated_at").defaultNow().onUpdateNow(),
+    deletedAt: timestamp("deleted_at"),
   },
 );
 
 export const PromotionalPrintingSizeEnum = mysqlTable(
   "promotional_printing_size_enum",
   {
-
+    id: int("id").autoincrement().primaryKey(),
+    statusName: varchar("printing_size", { length: 30 }),
+    createdAt: timestamp("created_at").defaultNow().notNull(),
+    updatedAt: timestamp("updated_at").defaultNow().onUpdateNow(),
+    deletedAt: timestamp("deleted_at"),
   },
 );
 
@@ -37,19 +38,60 @@ export const PromotionalPrintingSizeEnum = mysqlTable(
 export const PromotionalPrintingOrder = mysqlTable(
   "promotional_printing_order",
   {
-
+    id: int("id").autoincrement().primaryKey(),
+    clubId: int("club_id")
+      .notNull()
+      .references(() => Club.id),
+    studentId: int("student_id")
+      .notNull()
+      .references(() => Student.id),
+    studentPhoneNumber: varchar("student_phone_number", { length: 30 }),
+    promotionalPrintingOrderStatusEnum: int(
+      "promotional_printing_order_status_enum",
+    ).notNull(),
+    // .references(() => PromotionalPrintingOrderStatusEnum.id),
+    documentFileLink: text("document_file_link"),
+    isColorPrint: boolean("is_color_print").default(true).notNull(),
+    fitPrintSizeToPaper: boolean("fit_print_size_to_paper")
+      .default(true)
+      .notNull(),
+    requireMarginChopping: boolean("require_margin_chopping")
+      .default(false)
+      .notNull(),
+    desiredPickUpTime: datetime("desired_pick_up_time").notNull(),
+    pickUpAt: datetime("pick_up_at"),
+    createdAt: timestamp("created_at").defaultNow().notNull(),
+    updatedAt: timestamp("updated_at").defaultNow().onUpdateNow(),
+    deletedAt: timestamp("deleted_at"),
   },
   table => ({
-
+    promotionalPrintingOrderStatusEnumIdFk: index(
+      "pp_order_pp_order_status_enum_id_fk",
+    ).on(table.promotionalPrintingOrderStatusEnum),
   }),
 );
 
 export const PromotionalPrintingOrderSize = mysqlTable(
   "promotional_printing_order_size",
   {
-
+    id: int("id").autoincrement().primaryKey(),
+    promotionalPrintingOrderId: int("promotional_printing_order_id").notNull(),
+    // .references(() => PromotionalPrintingOrder.id),
+    promotionalPrintingSizeEnumId: int(
+      "promotional_printing_size_enum_id",
+    ).notNull(),
+    // .references(() => PromotionalPrintingSizeEnum.id),
+    numberOfPrints: int("number_of_prints").notNull(),
+    createdAt: timestamp("created_at").defaultNow().notNull(),
+    updatedAt: timestamp("updated_at").defaultNow().onUpdateNow(),
+    deletedAt: timestamp("deleted_at"),
   },
   table => ({
-
+    promotionalPrintingOrderIdFk: index("pp_order_size_pp_order_id_fk").on(
+      table.promotionalPrintingOrderId,
+    ),
+    promotionalPrintingSizeEnumIdFk: index(
+      "pp_order_size_pp_size_enum_id_fk",
+    ).on(table.promotionalPrintingSizeEnumId),
   }),
 );
