+++ 2024-summer-FE-seminar/packages/web/src/features/my/frame/_atomic/ChachaMyInfoFrame.tsx
@@ -4,12 +4,13 @@ import React, { useEffect, useState } from "react";
 
 import styled from "styled-components";
 
+import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import Card from "@sparcs-clubs/web/common/components/Card";
 import FoldableSectionTitle from "@sparcs-clubs/web/common/components/FoldableSectionTitle";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import PhoneInput from "@sparcs-clubs/web/common/components/Forms/PhoneInput";
 import Button from "@sparcs-clubs/web/common/components/Button";
-
+import useGetMyUser from "@sparcs-clubs/web/features/my/service/useGetMyUser";
 import chachaMockUpPhone from "@sparcs-clubs/web/features/my/user/_mock/chachaMockUpPhone";
 
 const ButtonWrapper = styled.div`
@@ -19,21 +20,35 @@ const ButtonWrapper = styled.div`
 
 const ChachaMyInfoFrame = () => {
   const [toggle, setToggle] = useState<boolean>(true);
-  const [mockPhone, setMockPhone] = useState<string>(
-    chachaMockUpPhone.phoneNumber,
-  );
-  const newMockPhone =
-    mockPhone.slice(0, 3) +
-    "-" +
-    mockPhone.slice(3, 7) +
-    "-" +
-    mockPhone.slice(7, 11);
-  const [phone, setPhone] = useState<string>(newMockPhone);
+  const { data, isError, isLoading } = useGetMyUser({});
+  // const [mockPhone, setMockPhone] = useState<string>("");
+  const [phone, setPhone] = useState<string>("");
+
+  // 데이터가 로드된 후 한 번만 초기 전화번호 설정
+  useEffect(() => {
+    if (data?.phoneNumber && phone === "") {
+      const formattedPhone =
+        data.phoneNumber.slice(0, 3) +
+        "-" +
+        data.phoneNumber.slice(3, 7) +
+        "-" +
+        data.phoneNumber.slice(7, 11);
+      setPhone(formattedPhone);
+      // setPhone(mockPhone);
+    }
+  }, [data]);
 
   const handlePhoneChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setPhone(event.target.value);
-  };
+    console.log("Event:", event);
+    console.log("Event Target:", event.target);
+    console.log("Event Target Value:", event.target?.value);
 
+    if (event.target?.value !== undefined) {
+      setPhone(event.target.value);
+    } else {
+      console.error("Event target value is undefined. Event:", event);
+    }
+  };
   return (
     <FlexWrapper direction="column" gap={40}>
       <FoldableSectionTitle
@@ -43,11 +58,13 @@ const ChachaMyInfoFrame = () => {
       />
       {toggle && (
         <Card padding="32px" gap={32}>
-          <PhoneInput
-            label="전화번호"
-            value={phone}
-            onChange={handlePhoneChange}
-          />
+          <AsyncBoundary isError={isError} isLoading={isLoading}>
+            <PhoneInput
+              label="전화번호"
+              value={phone}
+              onChange={handlePhoneChange}
+            />
+          </AsyncBoundary>
           <ButtonWrapper>
             <Button>저장</Button>
           </ButtonWrapper>
